Make actions into package and split up.
